// src/admin-app/src/app/invoices/constants/invoice-columns.enum.ts
export enum InvoiceColumn {
    issueDate = 'issueDate',
    dueDate = 'dueDate',
    company = 'company',
    country = 'country',
    itemCodes = 'itemCodes',
    priceWithoutVat = 'priceWithoutVat',
    VAT = 'VAT',
    price = 'price',
    currency = 'currency',
    invoiceCode = 'invoiceCode',
    paymentType = 'paymentType',
    paymentDate = 'paymentDate',
    paid = 'paid',
    unpaid = 'unpaid',
    author = 'author',
    note = 'note'
  }